+++ packages/vitest-environment-nuxt/src/runtime/mount.ts
@@ -1,5 +1,5 @@
-import { mount, VueWrapper, MountingOptions } from '@vue/test-utils'
-import { h, DefineComponent, Suspense, nextTick } from 'vue'
+import { mount, ComponentMountingOptions } from '@vue/test-utils'
+import { h, Suspense, nextTick, SetupContext, DefineComponent } from 'vue'
 import { defu } from 'defu'
 import type { RouteLocationRaw } from 'vue-router'
 
@@ -9,13 +9,14 @@ import { RouterLink } from './components/RouterLink'
 import NuxtRoot from '#build/root-component.mjs'
 import { useRouter } from '#imports'
 
-interface MountSuspendedOptions extends MountingOptions<any, any> {
+export type MountSuspendedOptions<T> = ComponentMountingOptions<T> & {
   route?: RouteLocationRaw
 }
 
-export async function mountSuspended<
-  T extends DefineComponent<any, any, any, any>
-> (component: T, options?: MountSuspendedOptions) {
+export async function mountSuspended<T>(
+  component: T,
+  options?: MountSuspendedOptions<T>
+) {
   const {
     props = {},
     attrs = {},
@@ -26,11 +27,25 @@ export async function mountSuspended<
 
   // @ts-expect-error untyped global __unctx__
   const vueApp = globalThis.__unctx__.get('nuxt-app').tryUse().vueApp
-  return new Promise<VueWrapper<InstanceType<T>>>(resolve => {
+  const { render, setup } = component as DefineComponent<any, any>
+
+  let setupContext: SetupContext
+  return new Promise<
+    ReturnType<
+      // @ts-expect-error letting uncertainty flow like water
+      typeof mount<T>
+    >
+  >(resolve => {
     const vm = mount(
       {
-        setup: NuxtRoot.setup,
-        render: () =>
+        setup: (props, ctx) => {
+          setupContext = ctx
+          return NuxtRoot.setup(props, {
+            ...ctx,
+            expose: () => {},
+          })
+        },
+        render: (renderContext: any) =>
           h(
             Suspense,
             { onResolve: () => nextTick().then(() => resolve(vm as any)) },
@@ -40,13 +55,29 @@ export async function mountSuspended<
                   async setup() {
                     const router = useRouter()
                     await router.replace(route)
-                    return () => h(component, { ...props, ...attrs }, slots)
+
+                    // Proxy top-level setup/render context so test wrapper resolves child component
+                    const clonedComponent = {
+                      ...component,
+                      render: render
+                        ? (_ctx: any, ...args: any[]) =>
+                            render(renderContext, ...args)
+                        : undefined,
+                      setup: setup
+                        ? (props: Record<string, any>) =>
+                            setup(props, setupContext)
+                        : undefined,
+                    }
+
+                    return () =>
+                      h(clonedComponent, { ...props, ...attrs }, slots)
                   },
                 }),
             }
           ),
       },
       defu(_options, {
+        slots,
         global: {
           config: {
             globalProperties: vueApp.config.globalProperties,