+++ playground/tests/nuxt/mock-nuxt-composable-2.spec.ts
@@ -1,0 +1,10 @@
+import { expect, it } from 'vitest'
+import { mockNuxtImport } from 'vitest-environment-nuxt/utils'
+
+mockNuxtImport('useHead', () => {
+  return () => true
+})
+
+it('should mock core nuxt composables', () => {
+  expect(useHead({})).toMatchInlineSnapshot('true')
+})